feat: refactor test folders
